requirements.txt
python-dotenv~=0.21.0
dateparser~=1.1.1
beautifulsoup4